Update add.php- the JavaScript variables and functions definitions
Please do not use var, use "let" instead, it is much safer since "var" value  is isolated  only inside functions, aka function-scope, whereas "let" is has a block-scope, means, let inside any block: loops, functions, anything... is safe from alteration by a new defined "let" variable with the same name.
As you noticed I also changed the functions' definitions to: 
a) and arrow function-  its easier to handle the-"this" issue that may be problematic if you are new to JavaScript: in a nutshell, arrow functions do not have thier own "this" so inside an arrow function "this" will refer to the Object the function is defined in- if it's not inside an object "this" will refer to the Window object of the browser. 
Plese read more about this very very important issue, even for experienced developers it is hard to grasp.
b) each arrow function is assined to a constant variable- since logically you do not plan to reassign the function's var/name to something alse.

Just a friendly reminder: If there are variables that are not suppose to change (or are defined inside functions and aren't beeng reassined in them....) you should make them constants.

Please MINDFULLY    apply these changes to the rest of your JavaScript code.